# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges/{id}/ranking:
    get:
      operationId: getChallengeRanking
      tags: [rankings]
      summary: Challenge leaderboard
      description: >
        Live standings for a challenge: rank, rank movement since the last
        scoring run, points, accuracy and exact/total predictions. The caller's
        own entry is highlighted via isCurrentUser and returned as me.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge ranking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingResponse"
        "403":
          description: Not permitted to view this challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rankings/global:
    get:
      operationId: getGlobalRanking
      tags: [rankings]
      summary: Global leaderboard
      description: >
        Platform-wide standings across all scored predictions. Each prediction
        is counted once. The caller's own entry is highlighted and returned as
        me even if outside the returned page.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Global ranking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingResponse"
  /rankings/competitions/{competitionSlug}:
    get:
      operationId: getCompetitionRanking
      tags: [rankings]
      summary: Per-competition leaderboard
      description: >
        Standings for a single competition's current season (or an explicit
        season key), derived live from scored predictions on that
        competition-season's matches. comingSoon is true when the competition
        has no current/upcoming season window yet, in which case the leaderboard
        is intentionally empty. The caller's own entry is returned as me even if
        outside the returned page.
      parameters:
        - name: competitionSlug
          in: path
          required: true
          schema:
            type: string
          description: Competition slug (e.g. eng.1, esp.1, ksa.1, fifa.world).
        - name: season
          in: query
          required: false
          schema:
            type: string
          description: >-
            Explicit season key. When omitted, the current/upcoming season is
            used (ended seasons are never surfaced implicitly).
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Competition ranking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionRankingResponse"
  /challenges/{id}/winning-probability:
    get:
      operationId: getWinningProbability
      tags: [rankings]
      summary: Caller's winning probability in a challenge
      description: >
        Heuristic estimate of the caller's chance to finish 1st, in the top 3
        and in the top 10, from current rank, points gap to the leader,
        accuracy and remaining matches. Gated by the winning_probability flag.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Winning probability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WinningProbability"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /challenges/{challengeId}/matches/{matchId}/impact:
    get:
      operationId: getRankingImpact
      tags: [rankings]
      summary: Estimated ranking impact of a live/upcoming match
      description: >
        For a match within a challenge, projects where the caller would rank if
        the current (live or predicted) result stands, e.g. moving from #8 to
        #4. Challenge-scoped to the caller.
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ranking impact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingImpact"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hall-of-fame/top-players:
    get:
      operationId: getTopPlayers
      tags: [gamification]
      summary: Global Top players across all challenges
      description: >
        Platform-wide Top-N players ranked by total points from all scored
        predictions, including each player's favourite team and the
        challenge(s) they take part in. Updates as matches are scored.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Top players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopPlayers"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Optional machine-readable error code for client-side branching
            (e.g. "owner_pool_full" when the host's shared participant pool
            is exhausted).
      required:
        - error
...
    FavoriteTeamRef:
      type: object
      description: Compact team reference for favourite team display.
      properties:
        id:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        flagUrl:
          type: ["string", "null"]
      required:
        - id
        - nameEn
        - nameAr
...
    RankingEntry:
      type: object
      properties:
        userId:
          type: string
        rank:
          type: integer
        previousRank:
          type: ["integer", "null"]
        rankMovement:
          type: integer
        displayName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        favoriteTeam:
          oneOf:
            - $ref: "#/components/schemas/FavoriteTeamRef"
            - type: "null"
        points:
          type: integer
        accuracy:
          type: ["number", "null"]
        exactPredictions:
          type: integer
        correctPredictions:
          type: integer
        totalPredictions:
          type: integer
        isCurrentUser:
          type: boolean
      required:
        - userId
        - rank
        - rankMovement
        - points
        - exactPredictions
        - correctPredictions
        - totalPredictions
        - isCurrentUser
    RankingResponse:
      type: object
      properties:
        scope:
          type: string
          enum: [challenge, global]
        challengeId:
          type: ["string", "null"]
        participantCount:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/RankingEntry"
        me:
          oneOf:
            - $ref: "#/components/schemas/RankingEntry"
            - type: "null"
      required:
        - scope
        - participantCount
        - entries
    CompetitionRankingResponse:
      type: object
      description: >-
        Per-competition leaderboard for the current season (or an explicit
        season key). Derived live from scored predictions on that
        competition-season's matches.
      properties:
        scope:
          type: string
          enum: [competition]
        competitionSlug:
          type: string
        season:
          type: ["string", "null"]
          description: The season key actually resolved (null when none exists yet).
        comingSoon:
          type: boolean
          description: >-
            True when the competition has no current/upcoming season window yet
            (only coming-soon shells or ended campaigns): the leaderboard is
            intentionally empty until fixtures publish.
        participantCount:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/RankingEntry"
        me:
          oneOf:
            - $ref: "#/components/schemas/RankingEntry"
            - type: "null"
      required:
        - scope
        - competitionSlug
        - comingSoon
        - participantCount
        - entries
...
    WinningProbability:
      type: object
      properties:
        challengeId:
          type: string
        rank:
          type: integer
        participants:
          type: integer
        points:
          type: integer
        leaderPoints:
          type: integer
        pointsGapToLead:
          type: integer
        remainingMatches:
          type: integer
        playedMatches:
          type: integer
        accuracy:
          type: ["number", "null"]
        firstPlacePct:
          type: number
        topThreePct:
          type: number
        topTenPct:
          type: number
      required:
        - challengeId
        - rank
        - participants
        - points
        - leaderPoints
        - pointsGapToLead
        - remainingMatches
        - playedMatches
        - firstPlacePct
        - topThreePct
        - topTenPct
    RankingImpact:
      type: object
      properties:
        challengeId:
          type: string
        matchId:
          type: string
        hasPrediction:
          type: boolean
        live:
          type: boolean
        currentRank:
          type: ["integer", "null"]
        projectedRank:
          type: ["integer", "null"]
        currentPoints:
          type: integer
        projectedPoints:
          type: integer
        pointsDelta:
          type: integer
      required:
        - challengeId
        - matchId
        - hasPrediction
        - live
        - currentPoints
        - projectedPoints
        - pointsDelta
...
    ChallengeRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    TopPlayerEntry:
      type: object
      properties:
        userId:
          type: string
        rank:
          type: integer
        rankMovement:
          type: integer
        displayName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        favoriteTeam:
          oneOf:
            - $ref: "#/components/schemas/FavoriteTeamRef"
            - type: "null"
        points:
          type: integer
        accuracy:
          type: ["number", "null"]
        exactPredictions:
          type: integer
        totalPredictions:
          type: integer
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeRef"
        isCurrentUser:
          type: boolean
      required:
        - userId
        - rank
        - rankMovement
        - points
        - exactPredictions
        - totalPredictions
        - challenges
        - isCurrentUser
    TopPlayers:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TopPlayerEntry"
        me:
          oneOf:
            - $ref: "#/components/schemas/TopPlayerEntry"
            - type: "null"
      required:
        - entries
...